More revisions to bindings API, thread safety improvements
Replaces single-event `poll_event/1` with multi-event `start_polling/1`
and renames `cancel_poll_event/0` to `stop_polling/0` to match. This
helps remove a lot of overhead (don't need a new thread for every single
event).

Updates to globals (used to track the termbox initialization and polling
thread) are now coordinated via a mutex to address race conditions that
could cause a crash. The locking logic adds a lot of noise to the code,
but makes it (hopefully) more thread-safe. This is a stopgap for now---I
plan to reimplement the bindings in Rust later.